updated components with login features and dummy data
